⚗️ Add tests for `charmcraft.yaml` parser
